Revert "Merge sol"
