+++ .data/datafile
@@ -1,7 +1,22 @@
-
-{"firstName":"","lastName":"lij","_id":"zYxIuNc5SahUR7B9"}
-{"$$deleted":true,"_id":"zYxIuNc5SahUR7B9"}
-{"firstName":"","lastName":"","_id":"t1A8j8jaUAPkqbqo"}
-{"firstName":"","lastName":"","_id":"RCKRAlncEtTyiTbH"}
-{"firstName":"","lastName":"","_id":"58z2ytBfZ5qWDcuv"}
-{"firstName":"","lastName":"","_id":"OFeD3txaKjBm4OEt"}
+{"firstName":"padsfjlk","lastName":"asdf","_id":"F0ztxfC5fMAO9iEq"}
+{"firstName":"sdf","lastName":"sadf","_id":"KJRk98uTCOWahoRj"}
+{"firstName":"","lastName":"","_id":"VMIOoI6h6CiBvi0X"}
+{"firstName":"fssadf","lastName":"dsfsdf","_id":"uPcxO053HvRO4VRQ"}
+{"firstName":"","lastName":"","_id":"1JPwKaJ2rc47I8xU"}
+{"firstName":"","lastName":"","_id":"xHL1LSyKLwTzPuTK"}
+{"firstName":"","lastName":"","_id":"je4hdOROGSpHKSai"}
+{"firstName":"","lastName":"","_id":"RPPiAnfj17SPpMhj"}
+{"firstName":"","lastName":"","_id":"TODBgJoaV2xN5NGJ"}
+{"$$deleted":true,"_id":"1JPwKaJ2rc47I8xU"}
+{"$$deleted":true,"_id":"F0ztxfC5fMAO9iEq"}
+{"$$deleted":true,"_id":"KJRk98uTCOWahoRj"}
+{"$$deleted":true,"_id":"RPPiAnfj17SPpMhj"}
+{"$$deleted":true,"_id":"TODBgJoaV2xN5NGJ"}
+{"$$deleted":true,"_id":"VMIOoI6h6CiBvi0X"}
+{"$$deleted":true,"_id":"je4hdOROGSpHKSai"}
+{"$$deleted":true,"_id":"uPcxO053HvRO4VRQ"}
+{"$$deleted":true,"_id":"xHL1LSyKLwTzPuTK"}
+{"firstName":"💻","lastName":"🌸","_id":"ucdZPLAzOQVOGMMg"}
+{"firstName":"+","lastName":"-","_id":"5ZSpF0GXKUCzFMIC"}
+{"$$deleted":true,"_id":"5ZSpF0GXKUCzFMIC"}
+{"$$deleted":true,"_id":"ucdZPLAzOQVOGMMg"}
