+++ public/assets-main/data/moonPhase_12_Nov-Dec.JSON
