+++ public/style.css
@@ -1,5 +1,8 @@
 /* dark: #0d0d0d rgb(13, 13, 13) */
 /* light: #eeeeee */
+/* green  #00da9b */
+@import url(https://static.mailerlite.com/assets/plugins/groot/modules/includes/groot_fonts/import.css?version=1607959);
+
 * {
   margin: 0;
   padding: 0;
@@ -34,10 +37,17 @@ header, section, aside, footer, canvas {
   /* border: 1px solid #ff0000; */
 }
 
-header, section, aside, footer {
+header, footer {
   display: block;
   margin: 0;
   padding: 1rem;
+  color: #eeeeee;
+}
+
+section, aside {
+  display: block;
+  margin: .5rem;
+  padding: 0;
   color: #eeeeee;
 }
 
@@ -50,6 +60,7 @@ main {
 .flex-container {
   display: flex;
   flex-direction: row;
+  justify-content: center;
 }
 
 .flex-wrap-container {
@@ -85,18 +96,19 @@ main {
   background: #03a9f4;
 }
 
-.hero {
-  /* background: #d22b1f; */
-}
+.hero {}
 
 .content {
   flex: 1;
-  /* background: #129a22; */
 }
 
-.sidebar {
-  flex: 0 1 300px;
-  border: 1px solid #000000;
+.sidebar-30 {
+  flex: 0 1 30%;
+}
+
+.sidebar-50 {
+  flex: 0 0 50%;
+  /* border: 1px solid #000000; */
   /* background: #673ab7; */
 }
 
@@ -139,6 +151,15 @@ main {
   color: #eeeeee;
 }
 
+/* ****** responsive font sizes ******* */
+.lg-font {
+  font-size: 2.3rem;
+}
+
+.md-font {
+  font-size: 1.2rem;
+}
+
 /* ******** Canvases ******** */
 #canvas-background {
   position: fixed;
@@ -149,13 +170,17 @@ main {
 }
 
 #p5Div, #threejsDiv {
-
   width: 100%;
   /* height: 400px; */
   border: .1 solid #eeeeee;
 }
 
-#threejsDiv{
+#threeContainer {
+  min-height: 66vh;
+
+}
+
+#threejsDiv {
   border-top: 2px solid #1f1f1f;
   border-right: 2px solid #1f1f1f;
   border-left: 2px solid #1f1f1f;
@@ -168,29 +193,27 @@ main {
   border-left: 2px solid #1f1f1f;
 }
 
-
 /* ************* Contributions *************  */
 .crosshatch {
-   /* width: 500px;
+  /* width: 500px;
    height: 200px; */
-   color: white;
-   background-image: linear-gradient(
-      45deg,
-      rgba(99, 99, 99,0.5) 25%,
+  color: white;
+  background-image: linear-gradient(45deg,
+      rgba(99, 99, 99, 0.5) 25%,
       transparent 25%,
       transparent 50%,
-      rgba(99, 99, 99,0.5) 50%,
-      rgba(99, 99, 99,0.5) 75%,
+      rgba(99, 99, 99, 0.5) 50%,
+      rgba(99, 99, 99, 0.5) 75%,
       transparent 75%,
-      transparent
-      );
-   background-size: 10px 3px;
-   border: 2px solid rgba(99, 99, 99,0.5);
+      transparent);
+  background-size: 10px 3px;
+  border: 2px solid rgba(99, 99, 99, 0.5);
 }
 
-.speakButton:hover{
+.speakButton:hover {
   background-color: #000;
 }
+
 .users li {
   display: inline;
 }
@@ -207,6 +230,7 @@ main {
 .contribution a:hover {
   color: #00ff00;
 }
+
 /* ************* Dropdown (moon phases) *************  */
 /* https://www.w3schools.com/css/css_dropdowns.asp */
 /* Style The Dropdown Button */
@@ -271,5 +295,24 @@ main {
 
   .flex-container {
     flex-direction: column;
+  }
+
+  .sidebar-50 {
+    /* flex: 0 1 100px; */
+    border: 1px solid #000000;
+  }
+
+  /* ****** responsive font sizes ******* */
+  .lg-font {
+    font-size: 1.2rem;
+  }
+
+  .md-font {
+    font-size: 1rem;
+  }
+
+  #threeContainer {
+    min-height: 30vh;
+
   }
 }
